+++ src/Snipcart.php
@@ -26,10 +26,13 @@ use workingconcept\snipcart\models\Settings;
 use workingconcept\snipcart\fields\ProductDetails;
 use workingconcept\snipcart\assetbundles\PluginSettingsAsset;
 use workingconcept\snipcart\events\RegisterShippingProvidersEvent;
+use workingconcept\snipcart\helpers\RouteHelper;
 use Craft;
 use craft\base\Plugin;
 use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterComponentTypesEvent;
+use craft\events\RegisterCacheOptionsEvent;
+use craft\utilities\ClearCaches;
 use craft\services\Fields;
 use craft\web\UrlManager;
 use craft\web\twig\variables\CraftVariable;
@@ -136,6 +139,23 @@ class Snipcart extends Plugin
             }
         );
 
+        Event::on(
+            ClearCaches::class,
+            ClearCaches::EVENT_REGISTER_CACHE_OPTIONS,
+            function (RegisterCacheOptionsEvent $event) {
+                $event->options = array_merge(
+                    $event->options,
+                    [
+                        [
+                            'key'    => Api::CACHE_TAG,
+                            'action' => Api::invalidateCache(),
+                            'label'  => Craft::t('snipcart', 'Snipcart API Cache'),
+                        ],
+                    ]
+                );
+            }
+        );
+
         if (Craft::$app->getRequest()->isCpRequest)
         {
             Event::on(
@@ -143,7 +163,10 @@ class Snipcart extends Plugin
                 UrlManager::EVENT_REGISTER_CP_URL_RULES,
                 function(RegisterUrlRulesEvent $event)
                 {
-                    $event->rules = array_merge($event->rules, $this->_cpRoutes());
+                    $event->rules = array_merge(
+                        $event->rules,
+                        RouteHelper::getCpRoutes()
+                    );
                 }
             );
         }
@@ -199,28 +222,6 @@ class Snipcart extends Plugin
 
     // Private Methods
     // =========================================================================
-
-    /**
-     * Define control panel routes.
-     * @return array
-     */
-    private function _cpRoutes(): array
-    {
-        return [
-            'snipcart' => 'snipcart/overview/index',
-            'snipcart/orders' => 'snipcart/orders/index',
-            'snipcart/order/<orderId>' => 'snipcart/orders/order-detail',
-            'snipcart/customers' => 'snipcart/customers/index',
-            'snipcart/customer/<customerId>' => 'snipcart/customers/customer-detail',
-            'snipcart/discounts' => 'snipcart/discounts/index',
-            'snipcart/discounts/new' => 'snipcart/discounts/new',
-            'snipcart/discount/<discountId>' => 'snipcart/discounts/discount-detail',
-            'snipcart/abandoned' => 'snipcart/carts/index',
-            'snipcart/abandoned/<cartId>' => 'snipcart/carts/detail',
-            'snipcart/subscriptions' => 'snipcart/subscriptions/index',
-            'snipcart/subscription/<subscriptionId>' => 'snipcart/subscriptions/detail',
-        ];
-    }
 
     /**
      * Instantiate Shipping providers and make each available in an indexed array.