+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 1.0.0-beta.20 - 2019-02-17
 ### Added
-- Added support for AbandonedCart::compatibilitySwitches.
+- Added support for additional AbandonedCart attributes.
 
 ## 1.0.0-beta.19 - 2019-02-16
 ### Added