+++ changelog.md
@@ -1,5 +1,9 @@
 # Snipcart Changelog
 
+## 1.0.0-beta.21 - 2019-02-17
+### Changed
+- Isolated API caches with TagDependency.
+
 ## 1.0.0-beta.20 - 2019-02-17
 ### Added
 - Added support for additional AbandonedCart attributes.